Fix instance envs endpoint (#134)
Response schema was defined with env vars being nested inside an instance object, this did not match the real API response.

Updates the response schema to change the types of `importedEnvs` and `readOnlyEnvs` from array to object to match the API.
Updates the type of `value` to acknowledge that the API may also return numbers in certain cases.